adding SecurityUtils dependency (refacto role with getRole)

+++ src/main/java/org/example/premier_projet_spring/controller/ProductController.java
@@ -3,12 +3,16 @@ package org.example.premier_projet_spring.controller;
 import jakarta.validation.Valid;
 import org.example.premier_projet_spring.dao.ProductDao;
 import org.example.premier_projet_spring.model.Product;
+import org.example.premier_projet_spring.model.Seller;
 import org.example.premier_projet_spring.model.State;
+import org.example.premier_projet_spring.security.AppUserDetails;
 import org.example.premier_projet_spring.security.IsClient;
 import org.example.premier_projet_spring.security.IsSeller;
+import org.example.premier_projet_spring.security.SecurityUtils;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
+import org.springframework.security.core.annotation.AuthenticationPrincipal;
 import org.springframework.web.bind.annotation.*;
 
 import java.util.List;
@@ -19,14 +23,15 @@ import java.util.Optional;
 public class ProductController {
 
     protected ProductDao productDao;
+    protected SecurityUtils securityUtils;
 
     @Autowired
-    public ProductController(ProductDao productDao) {
+    public ProductController(ProductDao productDao, SecurityUtils securityUtils) {
         this.productDao = productDao;
+        this.securityUtils = securityUtils;
     }
 
     @GetMapping("/product/{id}")
-
     @IsClient
     public ResponseEntity<Product> getProduct(@PathVariable Long id) {
 
@@ -39,15 +44,17 @@ public class ProductController {
 
 
     @GetMapping("/products")
-    @IsSeller
+    @IsClient
     public List<Product> getAll() {
         return productDao.findAll();
     }
 
     @PostMapping("/product")
+    @IsSeller
+    public ResponseEntity<Product> createProduct(@RequestBody @Valid Product product, @AuthenticationPrincipal AppUserDetails userDetails) {
 
 
-    public ResponseEntity<Product> createProduct(@RequestBody @Valid Product product) {
+        product.setCreator((Seller) userDetails.getUser());
 
         //si le produit n'a pas d'état il sera alors considéré comme neuf
         if (product.getState() == null) {
@@ -61,12 +68,19 @@ public class ProductController {
     }
 
     @DeleteMapping("product/{id}")
-    @IsClient
-    public ResponseEntity<Product> deleteProduct(@PathVariable Long id) {
+    @IsSeller
+    public ResponseEntity<Product> deleteProduct(@PathVariable Long id, @AuthenticationPrincipal AppUserDetails userDetails) {
 
         Optional<Product> optionalProduct = productDao.findById(id);
+
         if (optionalProduct.isEmpty()) {
             return new ResponseEntity<>(HttpStatus.NOT_FOUND);
+        }
+
+        String role = securityUtils.getRole(userDetails);
+
+        if (!role.equals("ROLE_CHIEF") || !optionalProduct.get().getCreator().equals(userDetails.getUser())) {
+            return new ResponseEntity<>(HttpStatus.FORBIDDEN);
         }
 
         productDao.deleteById(id);
@@ -75,15 +89,17 @@ public class ProductController {
 
     //Put change tout l'objet
     @PutMapping("/product/{id}")
+    @IsSeller
     // Patch change une partie de l'objet
 //    @PatchMapping("/product/{id}")
-    public ResponseEntity<Product> updateProduct(@PathVariable Long id, @RequestBody @Valid Product product) {
+    public ResponseEntity<Product> updateProduct(@PathVariable Long id, @RequestBody @Valid Product product, @AuthenticationPrincipal AppUserDetails userDetails) {
         Optional<Product> optionalProduct = productDao.findById(id);
         if (optionalProduct.isEmpty()) {
             return new ResponseEntity<>(HttpStatus.NOT_FOUND);
         }
 
         product.setId(id);
+        product.setCreator((Seller) userDetails.getUser());
         productDao.save(product);
         return new ResponseEntity<>(HttpStatus.NO_CONTENT);
     }
